Change CellNAPID from a short int to an int.  This will likely anger the RT.

+++ src/CCA/Components/MPM/ConstitutiveModel/MPMMaterial.h
@@ -109,7 +109,7 @@ WARNING
 
 
    particleIndex createParticles(
-                        CCVariable<short int>& cellNAPID,
+                        CCVariable<int>& cellNAPID,
                         const Patch*,
                         DataWarehouse* new_dw);
